HTML5 is now the default
