+++ src/wtforms/fields/__init__.py
@@ -3,6 +3,5 @@ from wtforms.fields.core import Field
 from wtforms.fields.core import Flags
 from wtforms.fields.core import Label
 from wtforms.fields.core import SelectFieldBase
-from wtforms.fields.html5 import *
 from wtforms.fields.simple import *
 from wtforms.utils import unset_value as _unset_value
